Move s3obj to files pkg
